+++ CHANGES.md
@@ -1,5 +1,6 @@
 [unreleased]
 * add filter `gu_parse_api_branches`
+* add filter `gu_display_repos`
 
 #### 0.7.0 / 2021-03-15 🎂
 * add filter `gu_post_api_response_body`